+++ src/main/java/org/nodens2k/tin/validation/BelgiumTinValidator.java
@@ -13,9 +13,9 @@ public final class BelgiumTinValidator extends AbstractCountryTinValidator {
     super("BE", "BEL", "056");
   }
 
-  @Contract(value = "null -> false", pure = true)
+  @Contract(value = "null,_ -> false", pure = true)
   @Override
-  public boolean isValid(String tin) {
+  public boolean isValid(String tin, TinType acceptedType) {
     tin = sanitise(tin, "BE");
     if (tin == null || !TIN_REGEX.matcher(tin).matches()) {
       return false;