+++ src/main/java/org/nodens2k/tin/validation/DenmarkTinValidator.java
@@ -20,7 +20,7 @@ public final class DenmarkTinValidator extends AbstractCountryTinValidator {
   }
 
   @Override
-  public boolean isValid(String tin) {
+  public boolean isValid(String tin, TinType acceptedType) {
     tin = sanitise(tin, "DK");
     if (tin == null || !TIN_REGEX.matcher(tin).matches()) {
       return false;