+++ src/main/java/org/nodens2k/tin/validation/EstoniaTinValidator.java
@@ -17,9 +17,9 @@ public final class EstoniaTinValidator extends AbstractCountryTinValidator {
     super("EE", "EST", "233");
   }
 
-  @Contract(value = "null -> false", pure = true)
+  @Contract(value = "null,_ -> false", pure = true)
   @Override
-  public boolean isValid(String tin) {
+  public boolean isValid(String tin, TinType acceptedType) {
     tin = sanitise(tin, "EE");
     if (tin == null || !TIN_REGEX.matcher(tin).matches()) {
       return false;