+++ src/main/java/org/nodens2k/tin/validation/FinlandTinValidator.java
@@ -20,7 +20,7 @@ public final class FinlandTinValidator extends AbstractCountryTinValidator {
   }
 
   @Override
-  public boolean isValid(String tin) {
+  public boolean isValid(String tin, TinType acceptedType) {
     tin = sanitise(tin, "FI");
     if (tin == null || !TIN_REGEX.matcher(tin).matches()) {
       return false;