+++ src/main/java/org/nodens2k/tin/validation/FranceTinValidator.java
@@ -13,9 +13,9 @@ public final class FranceTinValidator extends AbstractCountryTinValidator {
     super("FR", "FRA", "250");
   }
 
-  @Contract(value = "null -> false", pure = true)
+  @Contract(value = "null,_ -> false", pure = true)
   @Override
-  public boolean isValid(String tin) {
+  public boolean isValid(String tin, TinType acceptedType) {
     tin = sanitise(tin, "FR");
     if (tin == null || !TIN_REGEX.matcher(tin).matches()) {
       return false;