+++ src/main/java/org/nodens2k/tin/validation/GermanyTinValidator.java
@@ -16,9 +16,9 @@ public final class GermanyTinValidator extends AbstractCountryTinValidator {
     super("DE", "DEU", "276");
   }
 
-  @Contract(value = "null -> false", pure = true)
+  @Contract(value = "null,_ -> false", pure = true)
   @Override
-  public boolean isValid(String tin) {
+  public boolean isValid(String tin, TinType acceptedType) {
     tin = sanitise(tin, "DE");
     if (tin == null || !TIN_REGEX.matcher(tin).matches()) {
       return false;