+++ src/main/java/org/nodens2k/tin/validation/PortugalTinValidator.java
@@ -16,9 +16,9 @@ public final class PortugalTinValidator extends AbstractCountryTinValidator {
     super("PT", "PRT", "620");
   }
 
-  @Contract(value = "null -> false", pure = true)
+  @Contract(value = "null,_ -> false", pure = true)
   @Override
-  public boolean isValid(String tin) {
+  public boolean isValid(String tin, TinType acceptedType) {
     tin = sanitise(tin, "PT");
     if (tin == null || !TIN_REGEX.matcher(tin).matches()) {
       return false;