+++ src/main/java/org/nodens2k/tin/validation/SloveniaTinValidator.java
@@ -15,9 +15,9 @@ public final class SloveniaTinValidator extends AbstractCountryTinValidator {
     super("SI", "SVN", "705");
   }
 
-  @Contract(value = "null -> false", pure = true)
+  @Contract(value = "null,_ -> false", pure = true)
   @Override
-  public boolean isValid(String tin) {
+  public boolean isValid(String tin, TinType acceptedType) {
     tin = sanitise(tin, "SI");
     if (tin == null || !TIN_REGEX.matcher(tin).matches()) {
       return false;