+++ src/main/java/org/nodens2k/tin/validation/TinType.java
@@ -1,0 +1,19 @@
+package org.nodens2k.tin.validation;
+
+/**
+ * The different types of TIN to be validated.
+ */
+public enum TinType {
+  /**
+   * Any TIN valid document.
+   */
+  ANY,
+  /**
+   * TIN for a physical person.
+   */
+  PERSONAL,
+  /**
+   * TIN for a company.
+   */
+  COMPANY
+}