+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased (1.2.0-SNAPSHOT)
+
+### Added
+
+- Support for validating only company or personal TINs (only some countries)
+
 ## 1.1.0 - 2020-05-21
 
 ### Added